main function running on the Omega2+
I’ll include the ability to read prices for bitcoin first.

Future plans include:

* connecting to the Steem Blockchain via API to see voting power (and
possibly set it using buttons ?), mentions in the last 24Hours…

* Do an average price collection and/or apply trading price prediction
techniques directly on Omega2+ to show when it’s a good time to
buy/sell to reinvest.

* make it capable of maintaining an investment of 100$ in a crypto,
telling when to sell and when to buy in order to maintain it.