basic functionality implemented
I’ve made the Onion Omega2+ display the 3 Cryptocurrencies I’ve chosen,
using their stickers and then displaying the price.

1 line per crypto.

The display is initialised, the information shown for 10 seconds, and
then the OLED is deactivated, all in the same script.

Next step is to make functions allowing the user to choose a crypto
when running the script.

Also separate the initialisation, display, and shutdown into separate
scripts, then make a CRON job that initializes the display with a line
of “ XX.XX.XX HH.HH” corresponding to a date, then updates the display
every Y minutes/hours.